# backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todo API",
    "description": "A simple Todo application API built with FastAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/todos": {
      "get": {
        "tags": [
          "todos"
        ],
        "summary": "Get Todos",
        "description": "Get all todos for current user.\n\nArgs:\n    skip: Number of records to skip (pagination).\n    limit: Maximum number of records to return.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    List[Todo]: List of todo items.",
        "operationId": "get_todos_todos_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Groupid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  },
                  "title": "Response Get Todos Todos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todos"
        ],
        "summary": "Create Todo",
        "description": "Create a new todo.\n\nArgs:\n    todo: Todo data to create.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    Todo: The created todo item.",
        "operationId": "create_todo_todos_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{todoId}": {
      "get": {
        "tags": [
          "todos"
        ],
        "summary": "Read Todo",
        "description": "Get a specific todo by ID.\n\nArgs:\n    todoId: ID of the todo to retrieve.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    Todo: The requested todo item.\n\nRaises:\n    HTTPException: If todo not found or user doesn't have access.",
        "operationId": "read_todo_todos__todoId__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todoid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "todos"
        ],
        "summary": "Update Todo",
        "description": "Update an existing todo.\n\nArgs:\n    todoId: ID of the todo to update.\n    todo_update: Updated todo data.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    Todo: The updated todo item.\n\nRaises:\n    HTTPException: If todo not found or user is not the owner.",
        "operationId": "update_todo_todos__todoId__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todoid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "todos"
        ],
        "summary": "Delete Todo",
        "description": "Delete a todo.\n\nArgs:\n    todoId: ID of the todo to delete.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    dict: Success message.\n\nRaises:\n    HTTPException: If todo not found or user is not the owner.",
        "operationId": "delete_todo_todos__todoId__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todoid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Todo Todos  Todoid  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{todoId}/reorder": {
      "patch": {
        "tags": [
          "todos"
        ],
        "summary": "Reorder Todo",
        "description": "Update the index of a todo for drag-and-drop reordering.\n\nArgs:\n    todoId: ID of the todo to reorder.\n    reorder_data: Contains the new index.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    Todo: The reordered todo item.\n\nRaises:\n    HTTPException: If todo not found, user is not owner, or invalid index.",
        "operationId": "reorder_todo_todos__todoId__reorder_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todoid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Group": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "Group name"
          },
          "color": {
            "$ref": "#/components/schemas/GroupColor",
            "default": "blue"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "title": "Group",
        "description": "Schema for group API responses."
      },
      "GroupColor": {
        "type": "string",
        "enum": [
          "blue",
          "green",
          "red",
          "yellow",
          "purple",
          "pink",
          "orange",
          "teal",
          "indigo",
          "gray"
        ],
        "title": "GroupColor",
        "description": "Predefiniowane kolory dla grup - sp\u00f3jne z palet\u0105 UI."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Todo": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Title must not be empty"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "default": false
          },
          "index": {
            "type": "integer",
            "title": "Index",
            "default": 0
          },
          "groupId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupid"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "ownerEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owneremail"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Group"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "title": "Todo",
        "description": "Schema for todo API responses (includes id and timestamps)."
      },
      "TodoCreate": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Title must not be empty"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "default": false
          },
          "index": {
            "type": "integer",
            "title": "Index",
            "default": 0
          },
          "groupId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupid"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TodoCreate",
        "description": "Schema for creating a new todo."
      },
      "TodoUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed"
          },
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index"
          },
          "groupId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupid"
          }
        },
        "type": "object",
        "title": "TodoUpdate",
        "description": "Schema for updating an existing todo (all fields optional)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}